Collapse the headless transports into WooAssistantHeadless
Drops the URLSessionJetpackAIJWTClient wrapper in favor of a
WpComJetpackAIJWTProvider built directly inside WooAssistantHeadless.init
with a static mint closure. The cross-instance ProviderCache the wrapper
used to expose is gone - each harness instance mints its own JWT, which
is fine because smoke runs typically have one instance.

Inlines URLSessionWCRESTClient into the same file. The chat loop still
sees a WCRESTClient-conforming type; the file boundary served no
encapsulation purpose since the harness is the only consumer.

The Headless folder now ships a single file (WooAssistantHeadless.swift),
down from three.
